Refactoring

+++ DuckDB.NET.Data/DuckDBAppenderRow.cs
@@ -11,13 +11,13 @@ public class DuckDBAppenderRow
     private int columnIndex = 0;
     private readonly string qualifiedTableName;
     private readonly VectorDataWriterBase[] vectorWriters;
-    private readonly int rowIndex;
+    private readonly ulong rowIndex;
 
     internal DuckDBAppenderRow(string qualifiedTableName, VectorDataWriterBase[] vectorWriters, ulong rowIndex)
     {
         this.qualifiedTableName = qualifiedTableName;
         this.vectorWriters = vectorWriters;
-        this.rowIndex = (int)rowIndex;
+        this.rowIndex = rowIndex;
     }
 
     public void EndRow()
