+++ DuckDB.NET.Data/Internal/ScalarFunctionInfo.cs
@@ -6,10 +6,10 @@ using DuckDB.NET.Native;
 
 namespace DuckDB.NET.Data.Internal;
 
-class ScalarFunctionInfo(DuckDBLogicalType returnType, Action<VectorDataReaderBase[], VectorDataWriterBase, int> action) : IDisposable
+class ScalarFunctionInfo(DuckDBLogicalType returnType, Action<VectorDataReaderBase[], VectorDataWriterBase, ulong> action) : IDisposable
 {
     public DuckDBLogicalType ReturnType { get; } = returnType;
-    public Action<VectorDataReaderBase[], VectorDataWriterBase, int> Action { get; private set; } = action;
+    public Action<VectorDataReaderBase[], VectorDataWriterBase, ulong> Action { get; private set; } = action;
 
     public void Dispose()
     {
